refactor: Replace hand-rolled HTTP routing with Hono (#12)
Migrates all 50 HTTP routes from api.ts (1625 lines) and auth-routes.ts
integration to a single Hono router in routes.ts.

Changes:
- New routes.ts with Hono app handling all HTTP routes
- Auth middleware as Hono middleware (requireAuth)
- Auth routes delegated via app.all('/auth/*')
- Path params via :param instead of manual regex
- Centralized error handler via app.onError()
- WS upgrade stays in index.ts (before Hono, Bun needs server.upgrade)
- Deleted api.ts (fully replaced)

Hono 4.12.8 added as dependency. ~14KB runtime overhead.
All 50 routes preserved with identical behavior.
