--- src/concentrator/api.ts
@@ -1,1625 +1,0 @@
-/**
- * REST API for Concentrator
- * Provides endpoints for querying session data
- */
-
-import { randomUUID } from 'node:crypto'
-import { existsSync, mkdirSync, readdirSync, readFileSync, unlinkSync, writeFileSync } from 'node:fs'
-import { join } from 'node:path'
-import type { ListDirsResult, SendInput, Session, SpawnResult, TeamInfo } from '../shared/protocol'
-import { getGlobalSettings, updateGlobalSettings } from './global-settings'
-import { resolveInJail } from './path-jail'
-import {
-  deleteProjectSettings,
-  getAllProjectSettings,
-  getProjectSettings,
-  setProjectSettings,
-} from './project-settings'
-import { addSubscription, getSubscriptionCount, isPushConfigured, removeSubscription, sendPushToAll } from './push'
-import { getSessionOrder, moveSession, pinSession, setSessionOrder, unpinSession } from './session-order'
-import type { SessionStore } from './session-store'
-import { UI_HTML } from './ui'
-
-// Image registries
-// File registry: hash -> filesystem path (for [Image: source: /path] references)
-const fileRegistry = new Map<string, string>()
-// Blob registry: hash -> { bytes, mediaType, createdAt } (for inline base64 images from transcript)
-const blobRegistry = new Map<string, { bytes: Uint8Array; mediaType: string; createdAt: number }>()
-
-// Purge blobs older than 24 hours every hour
-const BLOB_MAX_AGE_MS = 24 * 60 * 60 * 1000
-setInterval(
-  () => {
-    const now = Date.now()
-    for (const [hash, entry] of blobRegistry) {
-      if (now - entry.createdAt > BLOB_MAX_AGE_MS) {
-        blobRegistry.delete(hash)
-      }
-    }
-  },
-  60 * 60 * 1000,
-)
-
-function hashString(input: string): string {
-  let hash = 0
-  for (let i = 0; i < input.length; i++) {
-    hash = (hash << 5) - hash + input.charCodeAt(i)
-    hash = hash & hash
-  }
-  return Math.abs(hash).toString(36)
-}
-
-function registerFilePath(path: string): string {
-  const hash = hashString(path)
-  fileRegistry.set(hash, path)
-  return hash
-}
-
-function registerBlob(data: string, mediaType: string): string {
-  // Hash the first 200 chars + length for speed (full base64 strings can be huge)
-  const key = `${data.length}:${data.slice(0, 200)}`
-  const hash = hashString(key)
-  if (!blobRegistry.has(hash)) {
-    const bytes = Buffer.from(data, 'base64')
-    blobRegistry.set(hash, { bytes: new Uint8Array(bytes), mediaType, createdAt: Date.now() })
-  }
-  return hash
-}
-
-function getImageSource(
-  hash: string,
-): { type: 'file'; path: string } | { type: 'blob'; bytes: Uint8Array; mediaType: string } | null {
-  const blob = blobRegistry.get(hash)
-  if (blob) return { type: 'blob', ...blob }
-  const path = fileRegistry.get(hash)
-  if (path) return { type: 'file', path }
-  return null
-}
-
-// Image extensions we recognize
-const IMAGE_EXTENSIONS = ['png', 'jpg', 'jpeg', 'gif', 'webp', 'avif', 'heic', 'svg']
-
-// Map media_type to extension
-function mediaTypeToExt(mediaType: string): string {
-  const map: Record<string, string> = {
-    'image/png': 'png',
-    'image/jpeg': 'jpg',
-    'image/gif': 'gif',
-    'image/webp': 'webp',
-    'image/avif': 'avif',
-    'image/heic': 'heic',
-    'image/svg+xml': 'svg',
-  }
-  return map[mediaType] || 'png'
-}
-
-/**
- * Process a transcript entry to find and register images.
- * Handles two sources:
- * 1. Inline base64 image blocks (type: "image", source.type: "base64") - always available
- * 2. File path references ([Image: source: /path/to/file.ext]) - needs filesystem access
- *
- * Returns the entry with `images` field added and base64 data stripped to save bandwidth.
- */
-function processImagesInEntry(entry: any): any {
-  const images: Array<{ hash: string; ext: string; url: string; originalPath: string }> = []
-  let modified = false
-
-  // 1. Extract inline base64 image blocks from message content
-  const content = entry?.message?.content
-  if (Array.isArray(content)) {
-    for (let i = 0; i < content.length; i++) {
-      const block = content[i]
-      if (
-        block?.type === 'image' &&
-        block?.source?.type === 'base64' &&
-        block?.source?.data &&
-        block?.source?.media_type
-      ) {
-        const mediaType = block.source.media_type as string
-        const ext = mediaTypeToExt(mediaType)
-        const hash = registerBlob(block.source.data, mediaType)
-        images.push({
-          hash,
-          ext,
-          url: `/file/${hash}.${ext}`,
-          originalPath: `inline:${mediaType}`,
-        })
-        // Replace the heavy base64 block with a lightweight placeholder
-        // so we don't send megabytes of base64 back to the dashboard
-        if (!modified) {
-          // Clone entry + content array on first modification
-          entry = { ...entry, message: { ...entry.message, content: [...content] } }
-          modified = true
-        }
-        entry.message.content[i] = {
-          type: 'text',
-          text: `[Image: ${hash}.${ext}]`,
-        }
-      }
-    }
-  }
-
-  // 2. Scan for file path references: [Image: source: /path/to/file.ext]
-  const imagePattern = /\[Image:\s*source:\s*([^\]]+)\]/gi
-
-  function scanText(value: any): void {
-    if (typeof value === 'string') {
-      let match
-      while ((match = imagePattern.exec(value)) !== null) {
-        const imagePath = match[1].trim()
-        const ext = imagePath.split('.').pop()?.toLowerCase() || 'png'
-        if (IMAGE_EXTENSIONS.includes(ext)) {
-          const hash = registerFilePath(imagePath)
-          // Don't add duplicate if we already have this hash from base64
-          if (!images.some(img => img.hash === hash)) {
-            images.push({
-              hash,
-              ext,
-              url: `/file/${hash}.${ext}`,
-              originalPath: imagePath,
-            })
-          }
-        }
-      }
-    } else if (Array.isArray(value)) {
-      value.forEach(scanText)
-    } else if (value && typeof value === 'object') {
-      Object.values(value).forEach(scanText)
-    }
-  }
-
-  scanText(entry)
-
-  if (images.length > 0) {
-    return { ...entry, images }
-  }
-  return entry
-}
-
-export interface ApiOptions {
-  sessionStore: SessionStore
-  webDir?: string
-  vapidPublicKey?: string
-  rclaudeSecret?: string
-  cacheDir?: string
-  serverStartTime?: number
-}
-
-// Build a map of embedded files for quick lookup
-type EmbeddedBlob = Blob & { name: string }
-const embeddedFiles = new Map<string, Blob>()
-const hasEmbeddedWeb = typeof Bun !== 'undefined' && (Bun.embeddedFiles as EmbeddedBlob[])?.length > 0
-
-if (hasEmbeddedWeb) {
-  for (const blob of Bun.embeddedFiles as EmbeddedBlob[]) {
-    // Remove hash from filename: "index-a1b2c3d4.html" -> "index.html"
-    const name = blob.name.replace(/-[a-f0-9]+\./, '.')
-    embeddedFiles.set(name, blob)
-    // Also map with lib/ prefix for assets
-    if (blob.name.startsWith('lib/') || blob.name.includes('/lib/')) {
-      const libPath = blob.name.includes('/lib/') ? blob.name.substring(blob.name.indexOf('/lib/') + 1) : blob.name
-      embeddedFiles.set(libPath, blob)
-    }
-  }
-}
-
-function getMimeType(path: string): string {
-  const ext = path.split('.').pop()?.toLowerCase()
-  const mimeTypes: Record<string, string> = {
-    html: 'text/html; charset=utf-8',
-    css: 'text/css; charset=utf-8',
-    js: 'application/javascript; charset=utf-8',
-    json: 'application/json; charset=utf-8',
-    svg: 'image/svg+xml',
-    png: 'image/png',
-    jpg: 'image/jpeg',
-    jpeg: 'image/jpeg',
-    gif: 'image/gif',
-    webp: 'image/webp',
-    avif: 'image/avif',
-    heic: 'image/heic',
-    ico: 'image/x-icon',
-    woff: 'font/woff',
-    woff2: 'font/woff2',
-    pdf: 'application/pdf',
-  }
-  return mimeTypes[ext || ''] || 'application/octet-stream'
-}
-
-interface SessionOverview {
-  id: string
-  cwd: string
-  model?: string
-  status: Session['status']
-  wrapperIds: string[]
-  startedAt: number
-  lastActivity: number
-  eventCount: number
-  activeSubagentCount: number
-  totalSubagentCount: number
-  team?: TeamInfo
-  lastEvent?: {
-    hookEvent: string
-    timestamp: number
-  }
-}
-
-/**
- * Create API request handler
- */
-export function createApiHandler(options: ApiOptions) {
-  const { sessionStore, webDir, vapidPublicKey, rclaudeSecret, cacheDir, serverStartTime = Date.now() } = options
-
-  function sessionToOverview(session: Session): SessionOverview {
-    const lastEvent = session.events[session.events.length - 1]
-    return {
-      id: session.id,
-      cwd: session.cwd,
-      model: session.model,
-      status: session.status,
-      wrapperIds: sessionStore.getWrapperIds(session.id),
-      startedAt: session.startedAt,
-      lastActivity: session.lastActivity,
-      eventCount: session.events.length,
-      activeSubagentCount: session.subagents.filter(a => a.status === 'running').length,
-      totalSubagentCount: session.subagents.length,
-      team: session.team,
-      lastEvent: lastEvent
-        ? {
-            hookEvent: lastEvent.hookEvent,
-            timestamp: lastEvent.timestamp,
-          }
-        : undefined,
-    }
-  }
-
-  return async function handleRequest(req: Request): Promise<Response> {
-    const url = new URL(req.url)
-    const path = url.pathname
-
-    // No CORS - same-origin only
-    if (req.method === 'OPTIONS') {
-      return new Response(null, { status: 204 })
-    }
-
-    // Serve embedded web dashboard if available
-    if (hasEmbeddedWeb && req.method === 'GET') {
-      // Serve index.html at root
-      if (path === '/' || path === '/index.html') {
-        const indexHtml = embeddedFiles.get('index.html')
-        if (indexHtml) {
-          return new Response(indexHtml, {
-            status: 200,
-            headers: { 'Content-Type': 'text/html; charset=utf-8' },
-          })
-        }
-      }
-
-      // Serve other embedded assets
-      const assetPath = path.startsWith('/') ? path.slice(1) : path
-      const asset = embeddedFiles.get(assetPath)
-      if (asset) {
-        return new Response(asset, {
-          status: 200,
-          headers: {
-            'Content-Type': getMimeType(assetPath),
-            'Cache-Control': assetPath.startsWith('lib/') ? 'public, max-age=31536000, immutable' : 'no-cache',
-          },
-        })
-      }
-
-      // SPA fallback - serve index.html for unknown paths (except API routes)
-      if (
-        !path.startsWith('/sessions') &&
-        !path.startsWith('/health') &&
-        !path.startsWith('/api') &&
-        !path.startsWith('/file')
-      ) {
-        const indexHtml = embeddedFiles.get('index.html')
-        if (indexHtml) {
-          return new Response(indexHtml, {
-            status: 200,
-            headers: { 'Content-Type': 'text/html; charset=utf-8' },
-          })
-        }
-      }
-    }
-
-    // Serve from webDir if specified
-    if (webDir && req.method === 'GET') {
-      const filePath = path === '/' ? '/index.html' : path
-      const fullPath = `${webDir}${filePath}`
-
-      // Path jail check - web files must stay within webDir
-      const safeWebPath = resolveInJail(fullPath)
-      if (!safeWebPath) {
-        // Fall through to other handlers instead of 403 (SPA routing)
-      } else
-        try {
-          const file = Bun.file(safeWebPath)
-          if (await file.exists()) {
-            const isAsset = filePath.startsWith('/assets/') || filePath.startsWith('/lib/')
-            return new Response(file, {
-              status: 200,
-              headers: {
-                'Content-Type': getMimeType(filePath),
-                'Cache-Control': isAsset ? 'public, max-age=31536000, immutable' : 'no-cache',
-              },
-            })
-          }
-
-          // SPA fallback - serve index.html for unknown paths (except API routes)
-          if (
-            !path.startsWith('/sessions') &&
-            !path.startsWith('/health') &&
-            !path.startsWith('/api') &&
-            !path.startsWith('/file')
-          ) {
-            const indexFile = Bun.file(`${webDir}/index.html`)
-            if (await indexFile.exists()) {
-              return new Response(indexFile, {
-                status: 200,
-                headers: { 'Content-Type': 'text/html; charset=utf-8' },
-              })
-            }
-          }
-        } catch {
-          // File not found, continue to other handlers
-        }
-    }
-
-    // Fallback UI at root (when no embedded web or webDir)
-    if ((path === '/' || path === '/ui') && req.method === 'GET' && !hasEmbeddedWeb && !webDir) {
-      return new Response(UI_HTML, {
-        status: 200,
-        headers: { 'Content-Type': 'text/html; charset=utf-8' },
-      })
-    }
-
-    // Health check
-    if (path === '/health') {
-      return new Response('ok', {
-        status: 200,
-        headers: { 'Content-Type': 'text/plain' },
-      })
-    }
-
-    // Server capabilities - tells dashboard what features are available
-    if (path === '/api/capabilities') {
-      return new Response(
-        JSON.stringify({
-          voice: !!process.env.DEEPGRAM_API_KEY,
-        }),
-        { status: 200, headers: { 'Content-Type': 'application/json' } },
-      )
-    }
-
-    // Serve registered images by hash: /file/{hash}.ext
-    const fileMatch = path.match(/^\/file\/([a-z0-9]+)(?:\.[a-z]+)?$/i)
-    if (fileMatch && req.method === 'GET') {
-      const hash = fileMatch[1]
-      const source = getImageSource(hash)
-
-      if (!source) {
-        return new Response(JSON.stringify({ error: 'Image not found' }), {
-          status: 404,
-          headers: { 'Content-Type': 'application/json' },
-        })
-      }
-
-      // Inline blob - serve directly from memory (no filesystem needed)
-      if (source.type === 'blob') {
-        return new Response(source.bytes, {
-          status: 200,
-          headers: {
-            'Content-Type': source.mediaType,
-            'Cache-Control': 'public, max-age=86400',
-          },
-        })
-      }
-
-      // File path - resolve through path jail
-      const safePath = resolveInJail(source.path)
-      if (!safePath) {
-        return new Response(JSON.stringify({ error: 'Access denied' }), {
-          status: 403,
-          headers: { 'Content-Type': 'application/json' },
-        })
-      }
-
-      try {
-        const file = Bun.file(safePath)
-        if (!(await file.exists())) {
-          return new Response(JSON.stringify({ error: 'File not found on disk' }), {
-            status: 404,
-            headers: { 'Content-Type': 'application/json' },
-          })
-        }
-
-        return new Response(file, {
-          status: 200,
-          headers: {
-            'Content-Type': getMimeType(safePath),
-            'Cache-Control': 'public, max-age=3600',
-          },
-        })
-      } catch (error) {
-        return new Response(JSON.stringify({ error: `Failed to serve file: ${error}` }), {
-          status: 500,
-          headers: { 'Content-Type': 'application/json' },
-        })
-      }
-    }
-
-    // List all sessions
-    if (path === '/sessions' && req.method === 'GET') {
-      const activeOnly = url.searchParams.get('active') === 'true'
-      const sessions = activeOnly ? sessionStore.getActiveSessions() : sessionStore.getAllSessions()
-
-      const summaries = sessions.map(sessionToOverview)
-
-      return new Response(JSON.stringify(summaries, null, 2), {
-        status: 200,
-        headers: { 'Content-Type': 'application/json' },
-      })
-    }
-
-    // Get session by ID
-    const sessionMatch = path.match(/^\/sessions\/([^/]+)$/)
-    if (sessionMatch && req.method === 'GET') {
-      const sessionId = sessionMatch[1]
-      const session = sessionStore.getSession(sessionId)
-
-      if (!session) {
-        return new Response(JSON.stringify({ error: 'Session not found' }), {
-          status: 404,
-          headers: { 'Content-Type': 'application/json' },
-        })
-      }
-
-      return new Response(JSON.stringify(sessionToOverview(session), null, 2), {
-        status: 200,
-        headers: { 'Content-Type': 'application/json' },
-      })
-    }
-
-    // Get session events
-    const eventsMatch = path.match(/^\/sessions\/([^/]+)\/events$/)
-    if (eventsMatch && req.method === 'GET') {
-      const sessionId = eventsMatch[1]
-      const limit = parseInt(url.searchParams.get('limit') || '0', 10)
-      const since = parseInt(url.searchParams.get('since') || '0', 10)
-      const events = sessionStore.getSessionEvents(sessionId, limit || undefined, since || undefined)
-
-      if (events.length === 0 && !sessionStore.getSession(sessionId)) {
-        return new Response(JSON.stringify({ error: 'Session not found' }), {
-          status: 404,
-          headers: { 'Content-Type': 'application/json' },
-        })
-      }
-
-      return new Response(JSON.stringify(events, null, 2), {
-        status: 200,
-        headers: { 'Content-Type': 'application/json' },
-      })
-    }
-
-    // Get session subagents
-    const subagentsMatch = path.match(/^\/sessions\/([^/]+)\/subagents$/)
-    if (subagentsMatch && req.method === 'GET') {
-      const sessionId = subagentsMatch[1]
-      const session = sessionStore.getSession(sessionId)
-
-      if (!session) {
-        return new Response(JSON.stringify({ error: 'Session not found' }), {
-          status: 404,
-          headers: { 'Content-Type': 'application/json' },
-        })
-      }
-
-      return new Response(JSON.stringify(session.subagents, null, 2), {
-        status: 200,
-        headers: { 'Content-Type': 'application/json' },
-      })
-    }
-
-    // Get session transcript (tail)
-    const transcriptMatch = path.match(/^\/sessions\/([^/]+)\/transcript$/)
-    if (transcriptMatch && req.method === 'GET') {
-      const sessionId = transcriptMatch[1]
-      const session = sessionStore.getSession(sessionId)
-
-      if (!session) {
-        return new Response(JSON.stringify({ error: 'Session not found' }), {
-          status: 404,
-          headers: { 'Content-Type': 'application/json' },
-        })
-      }
-
-      const limit = parseInt(url.searchParams.get('limit') || '20', 10)
-
-      // Serve from transcript cache (streamed from rclaude over WS)
-      if (!sessionStore.hasTranscriptCache(sessionId)) {
-        return new Response(JSON.stringify({ error: 'No transcript in cache (rclaude not streaming yet?)' }), {
-          status: 404,
-          headers: { 'Content-Type': 'application/json' },
-        })
-      }
-
-      const entries = sessionStore.getTranscriptEntries(sessionId, limit)
-      const processedEntries = entries.map((entry: any) => processImagesInEntry(entry))
-      return new Response(JSON.stringify(processedEntries, null, 2), {
-        status: 200,
-        headers: { 'Content-Type': 'application/json' },
-      })
-    }
-
-    // Get subagent transcript
-    const subagentTranscriptMatch = path.match(/^\/sessions\/([^/]+)\/subagents\/([^/]+)\/transcript$/)
-    if (subagentTranscriptMatch && req.method === 'GET') {
-      const sessionId = subagentTranscriptMatch[1]
-      const agentId = subagentTranscriptMatch[2]
-      const session = sessionStore.getSession(sessionId)
-
-      if (!session) {
-        return new Response(JSON.stringify({ error: 'Session not found' }), {
-          status: 404,
-          headers: { 'Content-Type': 'application/json' },
-        })
-      }
-
-      const limit = parseInt(url.searchParams.get('limit') || '100', 10)
-
-      // Serve from cache (streamed from rclaude over WS)
-      if (!sessionStore.hasSubagentTranscriptCache(sessionId, agentId)) {
-        return new Response(JSON.stringify({ error: 'No subagent transcript in cache (rclaude not streaming yet?)' }), {
-          status: 404,
-          headers: { 'Content-Type': 'application/json' },
-        })
-      }
-
-      const entries = sessionStore.getSubagentTranscriptEntries(sessionId, agentId, limit)
-      const processedEntries = entries.map((entry: any) => processImagesInEntry(entry))
-      return new Response(JSON.stringify(processedEntries, null, 2), {
-        status: 200,
-        headers: { 'Content-Type': 'application/json' },
-      })
-    }
-
-    // Send input to session
-    const inputMatch = path.match(/^\/sessions\/([^/]+)\/input$/)
-    if (inputMatch && req.method === 'POST') {
-      const sessionId = inputMatch[1]
-      const session = sessionStore.getSession(sessionId)
-
-      if (!session) {
-        return new Response(JSON.stringify({ error: 'Session not found' }), {
-          status: 404,
-          headers: { 'Content-Type': 'application/json' },
-        })
-      }
-
-      if (session.status === 'ended') {
-        return new Response(JSON.stringify({ error: 'Session has ended' }), {
-          status: 400,
-          headers: { 'Content-Type': 'application/json' },
-        })
-      }
-
-      const ws = sessionStore.getSessionSocket(sessionId)
-      if (!ws) {
-        return new Response(JSON.stringify({ error: 'Session not connected' }), {
-          status: 400,
-          headers: { 'Content-Type': 'application/json' },
-        })
-      }
-
-      try {
-        const body = (await req.json()) as { input: string; crDelay?: number }
-        if (!body.input || typeof body.input !== 'string') {
-          return new Response(JSON.stringify({ error: 'Missing input field' }), {
-            status: 400,
-            headers: { 'Content-Type': 'application/json' },
-          })
-        }
-
-        const inputMsg: SendInput = {
-          type: 'input',
-          sessionId,
-          input: body.input,
-          ...(typeof body.crDelay === 'number' && body.crDelay > 0 && { crDelay: body.crDelay }),
-        }
-        ws.send(JSON.stringify(inputMsg))
-
-        return new Response(JSON.stringify({ success: true }), {
-          status: 200,
-          headers: { 'Content-Type': 'application/json' },
-        })
-      } catch (error) {
-        return new Response(JSON.stringify({ error: `Failed to send input: ${error}` }), {
-          status: 500,
-          headers: { 'Content-Type': 'application/json' },
-        })
-      }
-    }
-
-    // POST /sessions/:id/revive - Revive an inactive session via host agent
-    const reviveMatch = path.match(/^\/sessions\/([^/]+)\/revive$/)
-    if (req.method === 'POST' && reviveMatch) {
-      const sessionId = reviveMatch[1]
-      const session = sessionStore.getSession(sessionId)
-      if (!session) {
-        return new Response(JSON.stringify({ error: 'Session not found' }), {
-          status: 404,
-          headers: { 'Content-Type': 'application/json' },
-        })
-      }
-
-      if (session.status === 'active') {
-        return new Response(JSON.stringify({ error: 'Session is already active' }), {
-          status: 400,
-          headers: { 'Content-Type': 'application/json' },
-        })
-      }
-
-      const agent = sessionStore.getAgent()
-      if (!agent) {
-        return new Response(JSON.stringify({ error: 'No host agent connected' }), {
-          status: 503,
-          headers: { 'Content-Type': 'application/json' },
-        })
-      }
-
-      try {
-        const wrapperId = randomUUID()
-        agent.send(
-          JSON.stringify({
-            type: 'revive',
-            sessionId,
-            cwd: session.cwd,
-            wrapperId,
-          }),
-        )
-
-        return new Response(JSON.stringify({ success: true, message: 'Revive command sent to agent', wrapperId }), {
-          status: 202,
-          headers: { 'Content-Type': 'application/json' },
-        })
-      } catch (error) {
-        return new Response(JSON.stringify({ error: `Failed to send revive: ${error}` }), {
-          status: 500,
-          headers: { 'Content-Type': 'application/json' },
-        })
-      }
-    }
-
-    // GET /api/agent/diag - Agent diagnostic log
-    if (req.method === 'GET' && path === '/api/agent/diag') {
-      const diag = {
-        connected: sessionStore.hasAgent(),
-        entries: sessionStore.getAgentDiag(),
-      }
-      return new Response(JSON.stringify(diag, null, 2), {
-        status: 200,
-        headers: { 'Content-Type': 'application/json' },
-      })
-    }
-
-    // POST /api/spawn - Spawn a new rclaude session at an arbitrary CWD
-    if (req.method === 'POST' && path === '/api/spawn') {
-      const agent = sessionStore.getAgent()
-      if (!agent) {
-        return new Response(JSON.stringify({ error: 'No host agent connected' }), {
-          status: 503,
-          headers: { 'Content-Type': 'application/json' },
-        })
-      }
-
-      try {
-        const body = (await req.json()) as { cwd: string; mkdir?: boolean }
-        if (!body.cwd || typeof body.cwd !== 'string') {
-          return new Response(JSON.stringify({ error: 'Missing cwd field' }), {
-            status: 400,
-            headers: { 'Content-Type': 'application/json' },
-          })
-        }
-
-        const requestId = randomUUID()
-        const wrapperId = randomUUID()
-
-        // Set up a one-shot response listener with timeout
-        const result = await new Promise<SpawnResult>((resolve, reject) => {
-          const timeout = setTimeout(() => {
-            sessionStore.removeSpawnListener(requestId)
-            reject(new Error('Spawn timed out (15s)'))
-          }, 15000)
-
-          sessionStore.addSpawnListener(requestId, (msg: SpawnResult) => {
-            clearTimeout(timeout)
-            resolve(msg)
-          })
-
-          agent.send(
-            JSON.stringify({
-              type: 'spawn',
-              requestId,
-              cwd: body.cwd,
-              wrapperId,
-              mkdir: body.mkdir || false,
-            }),
-          )
-        })
-
-        if (result.success) {
-          return new Response(JSON.stringify({ success: true, wrapperId, tmuxSession: result.tmuxSession }), {
-            status: 200,
-            headers: { 'Content-Type': 'application/json' },
-          })
-        }
-        return new Response(JSON.stringify({ error: result.error || 'Spawn failed' }), {
-          status: 500,
-          headers: { 'Content-Type': 'application/json' },
-        })
-      } catch (error) {
-        return new Response(JSON.stringify({ error: `Spawn failed: ${error}` }), {
-          status: 500,
-          headers: { 'Content-Type': 'application/json' },
-        })
-      }
-    }
-
-    // GET /api/dirs - List directories at a path (relayed to agent)
-    if (req.method === 'GET' && path === '/api/dirs') {
-      const agent = sessionStore.getAgent()
-      if (!agent) {
-        return new Response(JSON.stringify({ error: 'No host agent connected' }), {
-          status: 503,
-          headers: { 'Content-Type': 'application/json' },
-        })
-      }
-
-      const dirPath = url.searchParams.get('path') || '/'
-
-      try {
-        const requestId = randomUUID()
-
-        const result = await new Promise<ListDirsResult>((resolve, reject) => {
-          const timeout = setTimeout(() => {
-            sessionStore.removeDirListener(requestId)
-            reject(new Error('Directory listing timed out (5s)'))
-          }, 5000)
-
-          sessionStore.addDirListener(requestId, (msg: ListDirsResult) => {
-            clearTimeout(timeout)
-            resolve(msg)
-          })
-
-          agent.send(
-            JSON.stringify({
-              type: 'list_dirs',
-              requestId,
-              path: dirPath,
-            }),
-          )
-        })
-
-        if (result.error) {
-          return new Response(JSON.stringify({ error: result.error }), {
-            status: 400,
-            headers: { 'Content-Type': 'application/json' },
-          })
-        }
-
-        return new Response(JSON.stringify({ path: dirPath, dirs: result.dirs }), {
-          status: 200,
-          headers: { 'Content-Type': 'application/json' },
-        })
-      } catch (error) {
-        return new Response(JSON.stringify({ error: `${error}` }), {
-          status: 500,
-          headers: { 'Content-Type': 'application/json' },
-        })
-      }
-    }
-
-    // GET /sessions/:id/diag - Raw session diagnostics dump
-    const diagMatch = path.match(/^\/sessions\/([^/]+)\/diag$/)
-    if (diagMatch && req.method === 'GET') {
-      const sessionId = diagMatch[1]
-      const session = sessionStore.getSession(sessionId)
-      if (!session) {
-        return new Response(JSON.stringify({ error: 'Session not found' }), {
-          status: 404,
-          headers: { 'Content-Type': 'application/json' },
-        })
-      }
-      const transcriptCount = sessionStore.getTranscriptEntries(sessionId).length
-      const diag = {
-        id: sessionId,
-        cwd: session.cwd,
-        model: session.model,
-        status: session.status,
-        wrapperIds: sessionStore.getWrapperIds(sessionId),
-        capabilities: session.capabilities,
-        version: session.version,
-        buildTime: session.buildTime,
-        startedAt: session.startedAt,
-        lastActivity: session.lastActivity,
-        compacting: session.compacting,
-        compactedAt: session.compactedAt,
-        eventCount: session.events.length,
-        transcriptCacheEntries: transcriptCount,
-        subagents: session.subagents,
-        tasks: session.tasks,
-        bgTasks: session.bgTasks,
-        teammates: session.teammates,
-        team: session.team,
-        args: session.args,
-        diagLog: session.diagLog,
-      }
-      return new Response(JSON.stringify(diag, null, 2), {
-        status: 200,
-        headers: { 'Content-Type': 'application/json' },
-      })
-    }
-
-    // GET /sessions/:id/tasks - Get session task list
-    const tasksMatch = path.match(/^\/sessions\/([^/]+)\/tasks$/)
-    if (tasksMatch && req.method === 'GET') {
-      const sessionId = tasksMatch[1]
-      const session = sessionStore.getSession(sessionId)
-      if (!session) {
-        return new Response(JSON.stringify({ error: 'Session not found' }), {
-          status: 404,
-          headers: { 'Content-Type': 'application/json' },
-        })
-      }
-      return new Response(JSON.stringify({ tasks: session.tasks, archivedTasks: session.archivedTasks }, null, 2), {
-        status: 200,
-        headers: { 'Content-Type': 'application/json' },
-      })
-    }
-
-    // GET /agent/status - Check if host agent is connected
-    if (req.method === 'GET' && path === '/agent/status') {
-      return new Response(JSON.stringify({ connected: sessionStore.hasAgent() }), {
-        status: 200,
-        headers: { 'Content-Type': 'application/json' },
-      })
-    }
-
-    // POST /agent/quit - Tell the host agent to exit
-    if (req.method === 'POST' && path === '/agent/quit') {
-      const agent = sessionStore.getAgent()
-      if (!agent) {
-        return new Response(JSON.stringify({ error: 'No agent connected' }), {
-          status: 404,
-          headers: { 'Content-Type': 'application/json' },
-        })
-      }
-      agent.send(JSON.stringify({ type: 'quit', reason: 'Requested via API' }))
-      return new Response(JSON.stringify({ success: true }), {
-        status: 200,
-        headers: { 'Content-Type': 'application/json' },
-      })
-    }
-
-    // GET /api/push/vapid - return VAPID public key for push subscription
-    if (req.method === 'GET' && path === '/api/push/vapid') {
-      if (!vapidPublicKey) {
-        return new Response(JSON.stringify({ error: 'Push not configured' }), {
-          status: 503,
-          headers: { 'Content-Type': 'application/json' },
-        })
-      }
-      return new Response(JSON.stringify({ publicKey: vapidPublicKey, subscriptions: getSubscriptionCount() }), {
-        status: 200,
-        headers: { 'Content-Type': 'application/json' },
-      })
-    }
-
-    // POST /api/push/subscribe - register push subscription
-    if (req.method === 'POST' && path === '/api/push/subscribe') {
-      try {
-        const body = (await req.json()) as {
-          subscription: { endpoint: string; keys: { p256dh: string; auth: string } }
-        }
-        if (!body.subscription?.endpoint || !body.subscription?.keys) {
-          return new Response(JSON.stringify({ error: 'Invalid subscription' }), {
-            status: 400,
-            headers: { 'Content-Type': 'application/json' },
-          })
-        }
-        addSubscription(body.subscription, req.headers.get('user-agent') || undefined)
-        return new Response(JSON.stringify({ success: true, total: getSubscriptionCount() }), {
-          status: 200,
-          headers: { 'Content-Type': 'application/json' },
-        })
-      } catch {
-        return new Response(JSON.stringify({ error: 'Invalid JSON' }), {
-          status: 400,
-          headers: { 'Content-Type': 'application/json' },
-        })
-      }
-    }
-
-    // POST /api/push/unsubscribe - remove push subscription
-    if (req.method === 'POST' && path === '/api/push/unsubscribe') {
-      try {
-        const body = (await req.json()) as { endpoint: string }
-        if (!body.endpoint) {
-          return new Response(JSON.stringify({ error: 'Missing endpoint' }), {
-            status: 400,
-            headers: { 'Content-Type': 'application/json' },
-          })
-        }
-        removeSubscription(body.endpoint)
-        return new Response(JSON.stringify({ success: true }), {
-          status: 200,
-          headers: { 'Content-Type': 'application/json' },
-        })
-      } catch {
-        return new Response(JSON.stringify({ error: 'Invalid JSON' }), {
-          status: 400,
-          headers: { 'Content-Type': 'application/json' },
-        })
-      }
-    }
-
-    // POST /api/push/send - send push notification (auth'd with rclaude secret)
-    if (req.method === 'POST' && path === '/api/push/send') {
-      // Auth: require rclaude secret as Bearer token
-      const authHeader = req.headers.get('authorization')
-      const token = authHeader?.startsWith('Bearer ') ? authHeader.slice(7) : null
-      if (!rclaudeSecret || !token || token !== rclaudeSecret) {
-        return new Response(JSON.stringify({ error: 'Unauthorized' }), {
-          status: 401,
-          headers: { 'Content-Type': 'application/json' },
-        })
-      }
-
-      if (!isPushConfigured()) {
-        return new Response(JSON.stringify({ error: 'Push not configured (no VAPID keys)' }), {
-          status: 503,
-          headers: { 'Content-Type': 'application/json' },
-        })
-      }
-
-      try {
-        const rawBody = await req.text()
-        if (!rawBody) {
-          return new Response(JSON.stringify({ error: 'Empty request body' }), {
-            status: 400,
-            headers: { 'Content-Type': 'application/json' },
-          })
-        }
-        let body: { title: string; body: string; sessionId?: string; tag?: string }
-        try {
-          body = JSON.parse(rawBody)
-        } catch (_parseErr) {
-          return new Response(JSON.stringify({ error: 'Invalid JSON', received: rawBody.slice(0, 200) }), {
-            status: 400,
-            headers: { 'Content-Type': 'application/json' },
-          })
-        }
-        if (!body.title && !body.body) {
-          return new Response(JSON.stringify({ error: 'Need title or body' }), {
-            status: 400,
-            headers: { 'Content-Type': 'application/json' },
-          })
-        }
-        const result = await sendPushToAll({
-          title: body.title || 'rclaude',
-          body: body.body || '',
-          sessionId: body.sessionId,
-          tag: body.tag,
-        })
-        return new Response(JSON.stringify({ success: true, ...result }), {
-          status: 200,
-          headers: { 'Content-Type': 'application/json' },
-        })
-      } catch (error) {
-        return new Response(JSON.stringify({ error: `Send failed: ${error}` }), {
-          status: 500,
-          headers: { 'Content-Type': 'application/json' },
-        })
-      }
-    }
-
-    // POST /api/crash - record a dashboard crash report (public - no auth required)
-    if (req.method === 'POST' && path === '/api/crash') {
-      if (!cacheDir) {
-        return new Response(JSON.stringify({ error: 'No cache dir configured' }), {
-          status: 503,
-          headers: { 'Content-Type': 'application/json' },
-        })
-      }
-      try {
-        const body = await req.json()
-        const crashDir = join(cacheDir, 'crashes')
-        if (!existsSync(crashDir)) mkdirSync(crashDir, { recursive: true })
-
-        const ts = new Date().toISOString().replace(/[:.]/g, '-')
-        const file = join(crashDir, `crash-${ts}.json`)
-        const report = {
-          timestamp: new Date().toISOString(),
-          userAgent: req.headers.get('user-agent') || 'unknown',
-          ...(body as Record<string, unknown>),
-        }
-        writeFileSync(file, JSON.stringify(report, null, 2))
-
-        // Keep only latest 50 crash reports
-        const files = readdirSync(crashDir)
-          .filter(f => f.startsWith('crash-') && f.endsWith('.json'))
-          .sort()
-        if (files.length > 50) {
-          for (const old of files.slice(0, files.length - 50)) {
-            try {
-              unlinkSync(join(crashDir, old))
-            } catch {}
-          }
-        }
-
-        return new Response(JSON.stringify({ success: true, file: file.split('/').pop() }), {
-          status: 200,
-          headers: { 'Content-Type': 'application/json' },
-        })
-      } catch {
-        return new Response(JSON.stringify({ error: 'Invalid crash report' }), {
-          status: 400,
-          headers: { 'Content-Type': 'application/json' },
-        })
-      }
-    }
-
-    // GET /api/crashes - list recent crash reports
-    if (req.method === 'GET' && path === '/api/crashes') {
-      if (!cacheDir) {
-        return new Response(JSON.stringify([]), {
-          status: 200,
-          headers: { 'Content-Type': 'application/json' },
-        })
-      }
-      const crashDir = join(cacheDir, 'crashes')
-      if (!existsSync(crashDir)) {
-        return new Response(JSON.stringify([]), {
-          status: 200,
-          headers: { 'Content-Type': 'application/json' },
-        })
-      }
-      const files = readdirSync(crashDir)
-        .filter(f => f.startsWith('crash-') && f.endsWith('.json'))
-        .sort()
-        .reverse()
-        .slice(0, 20)
-      const reports = files.map(f => {
-        try {
-          return JSON.parse(readFileSync(join(crashDir, f), 'utf-8'))
-        } catch {
-          return { file: f, error: 'parse failed' }
-        }
-      })
-      return new Response(JSON.stringify(reports, null, 2), {
-        status: 200,
-        headers: { 'Content-Type': 'application/json' },
-      })
-    }
-
-    // GET /api/settings/projects - get all project settings
-    if (req.method === 'GET' && path === '/api/settings/projects') {
-      return new Response(JSON.stringify(getAllProjectSettings()), {
-        status: 200,
-        headers: { 'Content-Type': 'application/json' },
-      })
-    }
-
-    // POST /api/settings/projects - update project settings
-    if (req.method === 'POST' && path === '/api/settings/projects') {
-      try {
-        const body = (await req.json()) as { cwd: string; settings: { label?: string; icon?: string; color?: string } }
-        if (!body.cwd) {
-          return new Response(JSON.stringify({ error: 'Missing cwd' }), {
-            status: 400,
-            headers: { 'Content-Type': 'application/json' },
-          })
-        }
-        setProjectSettings(body.cwd, body.settings || {})
-        const allSettings = getAllProjectSettings()
-        // Broadcast to all dashboard subscribers
-        const json = JSON.stringify({ type: 'project_settings_updated', settings: allSettings })
-        for (const ws of sessionStore.getSubscribers()) {
-          try {
-            ws.send(json)
-          } catch {
-            /* dead socket */
-          }
-        }
-        return new Response(JSON.stringify({ success: true, settings: allSettings }), {
-          status: 200,
-          headers: { 'Content-Type': 'application/json' },
-        })
-      } catch (error) {
-        return new Response(JSON.stringify({ error: `Failed: ${error}` }), {
-          status: 400,
-          headers: { 'Content-Type': 'application/json' },
-        })
-      }
-    }
-
-    // DELETE /api/settings/projects - delete project settings
-    if (req.method === 'DELETE' && path === '/api/settings/projects') {
-      try {
-        const body = (await req.json()) as { cwd: string }
-        if (!body.cwd) {
-          return new Response(JSON.stringify({ error: 'Missing cwd' }), {
-            status: 400,
-            headers: { 'Content-Type': 'application/json' },
-          })
-        }
-        deleteProjectSettings(body.cwd)
-        const allSettings = getAllProjectSettings()
-        // Broadcast to all dashboard subscribers
-        const json = JSON.stringify({ type: 'project_settings_updated', settings: allSettings })
-        for (const ws of sessionStore.getSubscribers()) {
-          try {
-            ws.send(json)
-          } catch {
-            /* dead socket */
-          }
-        }
-        return new Response(JSON.stringify({ success: true, settings: allSettings }), {
-          status: 200,
-          headers: { 'Content-Type': 'application/json' },
-        })
-      } catch (error) {
-        return new Response(JSON.stringify({ error: `Failed: ${error}` }), {
-          status: 400,
-          headers: { 'Content-Type': 'application/json' },
-        })
-      }
-    }
-
-    // POST /api/settings/projects/generate-keyterms - auto-generate keyterms from project files
-    if (req.method === 'POST' && path === '/api/settings/projects/generate-keyterms') {
-      const openrouterKey = process.env.OPENROUTER_API_KEY
-      if (!openrouterKey) {
-        return new Response(JSON.stringify({ error: 'OPENROUTER_API_KEY not configured' }), {
-          status: 500,
-          headers: { 'Content-Type': 'application/json' },
-        })
-      }
-
-      try {
-        const body = (await req.json()) as { cwd: string }
-        if (!body.cwd) {
-          return new Response(JSON.stringify({ error: 'Missing cwd' }), {
-            status: 400,
-            headers: { 'Content-Type': 'application/json' },
-          })
-        }
-
-        // Find a connected wrapper session for this cwd to read files from
-        const allSessions = sessionStore.getAllSessions()
-        const sessionForCwd = allSessions.find(s => s.cwd === body.cwd && s.status === 'active')
-        const wrapperSocket = sessionForCwd ? sessionStore.getSessionSocket(sessionForCwd.id) : null
-        if (!wrapperSocket) {
-          return new Response(JSON.stringify({ error: 'No active session connected for this project' }), {
-            status: 503,
-            headers: { 'Content-Type': 'application/json' },
-          })
-        }
-
-        // Read project files via wrapper with timeout
-        const filesToRead = [
-          `${body.cwd}/CLAUDE.md`,
-          `${body.cwd}/.claude/CLAUDE.md`,
-          `${body.cwd}/package.json`,
-          `${body.cwd}/README.md`,
-        ]
-
-        const fileContents: string[] = []
-        for (const filePath of filesToRead) {
-          try {
-            const content = await new Promise<string | null>((resolve, reject) => {
-              const requestId = randomUUID()
-              const timeout = setTimeout(() => {
-                sessionStore.removeFileListener(requestId)
-                reject(new Error(`File read timed out (5s): ${filePath}`))
-              }, 5000)
-
-              sessionStore.addFileListener(requestId, (msg: { data?: string; error?: string }) => {
-                clearTimeout(timeout)
-                if (msg.error || !msg.data) {
-                  resolve(null)
-                } else {
-                  // file_response returns base64 data
-                  resolve(Buffer.from(msg.data, 'base64').toString('utf-8'))
-                }
-              })
-
-              wrapperSocket.send(JSON.stringify({ type: 'file_request', requestId, path: filePath }))
-            })
-            if (content) {
-              fileContents.push(`--- ${filePath} ---\n${content.slice(0, 10000)}`)
-            }
-          } catch {
-            // File not found or timeout - skip
-          }
-        }
-
-        if (fileContents.length === 0) {
-          return new Response(
-            JSON.stringify({ error: 'No project files found (CLAUDE.md, package.json, README.md)' }),
-            {
-              status: 404,
-              headers: { 'Content-Type': 'application/json' },
-            },
-          )
-        }
-
-        console.log(`[keyterms] Generating keyterms for ${body.cwd} from ${fileContents.length} files`)
-
-        // Extract keyterms via Haiku
-        const llmRes = await fetch('https://openrouter.ai/api/v1/chat/completions', {
-          method: 'POST',
-          headers: {
-            Authorization: `Bearer ${openrouterKey}`,
-            'Content-Type': 'application/json',
-          },
-          body: JSON.stringify({
-            model: 'anthropic/claude-haiku-4-5-20251001',
-            messages: [
-              {
-                role: 'system',
-                content: `Extract domain-specific terms from these project files for voice transcription keyword boosting. Focus on:
-- Project names, tool names, library names
-- Technical terms specific to this project
-- Abbreviations, acronyms, unusual spellings
-- Brand names, product names
-- Any term a speech-to-text engine would likely misspell
-
-Output a JSON array of strings. Each string should be the correct spelling of one term. Include 10-30 terms, most important first. Only output the JSON array, nothing else.`,
-              },
-              { role: 'user', content: fileContents.join('\n\n') },
-            ],
-            max_tokens: 1024,
-          }),
-        })
-
-        if (!llmRes.ok) {
-          const err = await llmRes.text().catch(() => '')
-          console.error(`[keyterms] LLM failed: ${llmRes.status} ${err.slice(0, 500)}`)
-          return new Response(JSON.stringify({ error: 'Failed to generate keyterms' }), {
-            status: 500,
-            headers: { 'Content-Type': 'application/json' },
-          })
-        }
-
-        const llmData = (await llmRes.json()) as { choices?: Array<{ message?: { content?: string } }> }
-        const raw = llmData.choices?.[0]?.message?.content?.trim() || '[]'
-        let keyterms: string[]
-        try {
-          // Parse JSON, handling potential markdown code fences
-          const cleaned = raw.replace(/^```(?:json)?\s*/, '').replace(/\s*```$/, '')
-          keyterms = JSON.parse(cleaned)
-          if (!Array.isArray(keyterms)) throw new Error('Not an array')
-          keyterms = keyterms.filter(t => typeof t === 'string' && t.trim()).map(t => t.trim())
-        } catch {
-          console.error(`[keyterms] Failed to parse LLM output: ${raw.slice(0, 200)}`)
-          return new Response(JSON.stringify({ error: 'Failed to parse keyterms from LLM' }), {
-            status: 500,
-            headers: { 'Content-Type': 'application/json' },
-          })
-        }
-
-        console.log(`[keyterms] Generated ${keyterms.length} keyterms: ${keyterms.join(', ')}`)
-
-        // Save to project settings
-        setProjectSettings(body.cwd, { keyterms })
-
-        return new Response(JSON.stringify({ keyterms, settings: getAllProjectSettings() }), {
-          status: 200,
-          headers: { 'Content-Type': 'application/json' },
-        })
-      } catch (error) {
-        console.error(`[keyterms] Error: ${error}`)
-        return new Response(JSON.stringify({ error: `Failed: ${error}` }), {
-          status: 500,
-          headers: { 'Content-Type': 'application/json' },
-        })
-      }
-    }
-
-    // GET /api/settings - get global settings + schema
-    if (req.method === 'GET' && path === '/api/settings') {
-      return new Response(JSON.stringify(getGlobalSettings()), {
-        status: 200,
-        headers: { 'Content-Type': 'application/json' },
-      })
-    }
-
-    // POST /api/settings - update global settings (soft-fail validation)
-    if (req.method === 'POST' && path === '/api/settings') {
-      try {
-        const body = await req.json()
-        const result = updateGlobalSettings(body)
-        // Broadcast settings_updated to all dashboard subscribers
-        const json = JSON.stringify({ type: 'settings_updated', settings: result.settings })
-        for (const ws of sessionStore.getSubscribers()) {
-          try {
-            ws.send(json)
-          } catch {
-            /* dead socket */
-          }
-        }
-        return new Response(JSON.stringify(result), {
-          status: 200,
-          headers: { 'Content-Type': 'application/json' },
-        })
-      } catch (error) {
-        return new Response(JSON.stringify({ error: `Failed: ${error}` }), {
-          status: 400,
-          headers: { 'Content-Type': 'application/json' },
-        })
-      }
-    }
-
-    // POST /api/files - upload a file, returns URL for Claude
-    if (req.method === 'POST' && path === '/api/files') {
-      try {
-        const contentType = req.headers.get('content-type') || ''
-        let bytes: Uint8Array
-        let mediaType: string
-        let filename = 'image'
-
-        if (contentType.includes('multipart/form-data')) {
-          const formData = await req.formData()
-          const file = formData.get('file') as File | null
-          if (!file) {
-            return new Response(JSON.stringify({ error: 'No file in form data' }), {
-              status: 400,
-              headers: { 'Content-Type': 'application/json' },
-            })
-          }
-          bytes = new Uint8Array(await file.arrayBuffer())
-          mediaType = file.type || 'image/png'
-          filename = file.name || 'image'
-        } else {
-          // Raw binary upload with Content-Type header
-          bytes = new Uint8Array(await req.arrayBuffer())
-          mediaType = contentType.split(';')[0] || 'image/png'
-          const ext = mediaType.split('/')[1]?.replace('jpeg', 'jpg') || 'png'
-          filename = `paste.${ext}`
-        }
-
-        // Hash and register
-        const key = `${bytes.length}:${Array.from(bytes.slice(0, 200)).join(',')}`
-        const hash = hashString(key)
-        if (!blobRegistry.has(hash)) {
-          blobRegistry.set(hash, { bytes, mediaType, createdAt: Date.now() })
-        }
-
-        const ext = mediaType.split('/')[1]?.replace('jpeg', 'jpg') || 'png'
-        const filePath = `/file/${hash}.${ext}`
-
-        // Build absolute URL from request Host header so Claude can fetch it
-        const host = req.headers.get('host') || 'localhost:9999'
-        const proto = req.headers.get('x-forwarded-proto') || (host.includes('localhost') ? 'http' : 'https')
-        const url = `${proto}://${host}${filePath}`
-
-        return new Response(JSON.stringify({ hash, url, filename, mediaType }), {
-          status: 200,
-          headers: { 'Content-Type': 'application/json' },
-        })
-      } catch (error) {
-        return new Response(JSON.stringify({ error: `Upload failed: ${error}` }), {
-          status: 500,
-          headers: { 'Content-Type': 'application/json' },
-        })
-      }
-    }
-
-    // GET /api/session-order - get session pin/organize metadata
-    if (req.method === 'GET' && path === '/api/session-order') {
-      return new Response(JSON.stringify(getSessionOrder()), {
-        headers: { 'Content-Type': 'application/json' },
-      })
-    }
-
-    // POST /api/session-order - update session order (full replace or action)
-    if (req.method === 'POST' && path === '/api/session-order') {
-      try {
-        const body = (await req.json()) as {
-          action?: 'pin' | 'unpin' | 'move' | 'set'
-          cwd?: string
-          toIndex?: number
-          organized?: Array<{ cwd: string; group?: string }>
-        }
-
-        if (body.action === 'pin' && body.cwd) {
-          pinSession(body.cwd)
-        } else if (body.action === 'unpin' && body.cwd) {
-          unpinSession(body.cwd)
-        } else if (body.action === 'move' && body.cwd && body.toIndex !== undefined) {
-          moveSession(body.cwd, body.toIndex)
-        } else if (body.action === 'set' && body.organized) {
-          setSessionOrder({ organized: body.organized })
-        } else {
-          return new Response(JSON.stringify({ error: 'Invalid action' }), {
-            status: 400,
-            headers: { 'Content-Type': 'application/json' },
-          })
-        }
-
-        const order = getSessionOrder()
-        // Broadcast to all dashboard subscribers
-        const json = JSON.stringify({ type: 'session_order_updated', order })
-        for (const ws of sessionStore.getSubscribers()) {
-          try {
-            ws.send(json)
-          } catch {
-            /* dead socket */
-          }
-        }
-        return new Response(JSON.stringify({ success: true, order }), {
-          status: 200,
-          headers: { 'Content-Type': 'application/json' },
-        })
-      } catch (error) {
-        return new Response(JSON.stringify({ error: `Failed: ${error}` }), {
-          status: 400,
-          headers: { 'Content-Type': 'application/json' },
-        })
-      }
-    }
-
-    // POST /api/transcribe - transcribe audio via Deepgram Nova-3
-    if (path === '/api/transcribe' && req.method === 'POST') {
-      const deepgramKey = process.env.DEEPGRAM_API_KEY
-      if (!deepgramKey) {
-        console.error('[transcribe] DEEPGRAM_API_KEY not configured - cannot transcribe')
-        return new Response(JSON.stringify({ error: 'DEEPGRAM_API_KEY not configured' }), {
-          status: 500,
-          headers: { 'Content-Type': 'application/json' },
-        })
-      }
-
-      try {
-        const body = (await req.json()) as { audioUrl?: string; sessionId?: string }
-        if (!body.audioUrl) {
-          console.error('[transcribe] Missing audioUrl in request body')
-          return new Response(JSON.stringify({ error: 'audioUrl required' }), {
-            status: 400,
-            headers: { 'Content-Type': 'application/json' },
-          })
-        }
-
-        console.log(`[transcribe] Fetching audio: ${body.audioUrl}`)
-
-        // Fetch the audio file
-        const audioRes = await fetch(body.audioUrl)
-        if (!audioRes.ok) throw new Error(`Failed to fetch audio: ${audioRes.status}`)
-        const audioBytes = new Uint8Array(await audioRes.arrayBuffer())
-        const contentType = audioRes.headers.get('content-type') || 'audio/webm'
-        console.log(`[transcribe] Audio: ${audioBytes.byteLength} bytes, type: ${contentType}`)
-
-        // Build keyterms from project settings (session -> cwd -> project keyterms)
-        const keyterms: string[] = []
-        if (body.sessionId) {
-          const session = sessionStore.getSession(body.sessionId)
-          if (session?.cwd) {
-            const projSettings = getProjectSettings(session.cwd)
-            if (projSettings?.keyterms?.length) {
-              keyterms.push(...projSettings.keyterms)
-              console.log(`[transcribe] Project keyterms for ${session.cwd}: ${projSettings.keyterms.join(', ')}`)
-            }
-          }
-        }
-        const params = new URLSearchParams({
-          model: 'nova-3',
-          smart_format: 'true',
-          punctuate: 'true',
-          filler_words: 'false', // strip um, uh, etc.
-          diarize: 'false',
-          language: 'en',
-        })
-        // Add keyterms individually (Nova-3 uses 'keyterm' not 'keywords')
-        for (const kt of keyterms) {
-          params.append('keyterm', kt)
-        }
-
-        console.log('[transcribe] Calling Deepgram Nova-3...')
-        const dgRes = await fetch(`https://api.deepgram.com/v1/listen?${params}`, {
-          method: 'POST',
-          headers: {
-            Authorization: `Token ${deepgramKey}`,
-            'Content-Type': contentType,
-          },
-          body: audioBytes,
-        })
-
-        if (!dgRes.ok) {
-          const err = await dgRes.text()
-          console.error(`[transcribe] Deepgram failed: ${dgRes.status} ${err.slice(0, 500)}`)
-          throw new Error(`Deepgram transcription failed: ${dgRes.status}`)
-        }
-
-        const dgData = (await dgRes.json()) as {
-          results?: {
-            channels?: Array<{
-              alternatives?: Array<{ transcript?: string }>
-            }>
-          }
-        }
-        const rawText = dgData.results?.channels?.[0]?.alternatives?.[0]?.transcript?.trim() || ''
-        console.log(`[transcribe] Result: "${rawText.slice(0, 200)}"${rawText.length > 200 ? '...' : ''}`)
-
-        if (!rawText.trim()) {
-          console.log('[transcribe] Empty transcription, returning empty')
-          return new Response(JSON.stringify({ raw: '', refined: '' }), {
-            status: 200,
-            headers: { 'Content-Type': 'application/json' },
-          })
-        }
-
-        // Deepgram with keyterms is good enough to skip Haiku refinement for now
-        // raw and refined are the same - no second LLM pass needed
-        return new Response(JSON.stringify({ raw: rawText, refined: rawText }), {
-          status: 200,
-          headers: { 'Content-Type': 'application/json' },
-        })
-      } catch (error) {
-        console.error('[transcribe] Pipeline error:', error)
-        return new Response(JSON.stringify({ error: `Transcription failed: ${error}` }), {
-          status: 500,
-          headers: { 'Content-Type': 'application/json' },
-        })
-      }
-    }
-
-    // GET /api/stats - server metrics
-    if (req.method === 'GET' && path === '/api/stats') {
-      const allSessions = sessionStore.getAllSessions()
-      let active = 0
-      let idle = 0
-      let ended = 0
-      for (const s of allSessions) {
-        if (s.status === 'active') active++
-        else if (s.status === 'idle') idle++
-        else ended++
-      }
-
-      const diag = sessionStore.getSubscriptionsDiag()
-      const traffic = sessionStore.getTrafficStats()
-
-      return new Response(
-        JSON.stringify({
-          uptime: Math.round((Date.now() - serverStartTime) / 1000),
-          sessions: { total: allSessions.length, active, idle, ended },
-          connections: {
-            total: diag.summary.totalSubscribers,
-            legacy: diag.summary.legacySubscribers,
-            v2: diag.summary.v2Subscribers,
-          },
-          traffic,
-          channels: diag.summary.channelCounts,
-        }),
-        { headers: { 'Content-Type': 'application/json' } },
-      )
-    }
-
-    // GET /api/subscriptions - diagnostic view of channel subscriptions
-    if (req.method === 'GET' && path === '/api/subscriptions') {
-      const diag = sessionStore.getSubscriptionsDiag()
-      return new Response(JSON.stringify(diag, null, 2), {
-        headers: { 'Content-Type': 'application/json' },
-      })
-    }
-
-    return new Response(JSON.stringify({ error: 'Not found' }), {
-      status: 404,
-      headers: { 'Content-Type': 'application/json' },
-    })
-  }
-}
